api/requirements.txt
affine==2.4.0
aiohttp==3.8.6
aioitertools==0.11.0
aiosignal==1.3.1
amqp==5.1.1
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.5.2
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.1.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.12.1
boto3==1.29.1
botocore==1.32.1
bs4==0.0.1
celery==5.3.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
cryptography==38.0.4
dataclasses-json==0.6.3
decouple==0.0.7
distlib==0.3.8
distro==1.8.0
Django==3.2.20
django-appconf==1.0.5
django-compressor==4.1
django-dotenv==1.4.2
django-formtools==2.4
django-otp==1.1.4
django-otp-yubikey==1.0.1
django-phonenumber-field==6.4.0
django-storages==1.13.2
django-two-factor-auth==1.14.0
ecdsa==0.18.0
environ==1.0
envs==1.4
et-xmlfile==1.1.0
exceptiongroup==1.2.0
fastapi==0.109.0
filelock==3.12.2
fiona==1.9.5
frozenlist==1.4.0
fsspec==2023.10.0
future==0.18.3
GDAL==3.7.2
gdown==4.7.1
geopandas==0.13.2
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.25.1
idna==3.6
importlib-metadata==6.7.0
iniconfig==2.0.0
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.1
langchain==0.0.354
langchain-community==0.0.8
langchain-core==0.1.6
langsmith==0.0.77
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.20.1
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.24.4
openai==1.2.4
openpyxl==3.1.2
packaging==23.2
pandas==2.0.3
pathspec==0.12.1
phonenumbers==8.13.2
Pillow==9.3.0
pipenv==2023.11.15
platformdirs==4.1.0
pluggy==1.3.0
prompt-toolkit==3.0.39
psycopg2==2.9.9
psycopg2-binary==2.9.7
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.3.1
pydantic==2.5.3
pydantic_core==2.14.6
pyotp==2.8.0
pyparsing==3.1.1
pypng==0.20220715.0
pyproj==3.5.0
PySocks==1.7.1
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-mock==3.12.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-jose-cryptodome==1.3.2
pytz==2022.6
PyYAML==6.0.1
qrcode==7.3.1
rasterio==1.3.9
rcssmin==1.1.0
requests==2.28.1
rjsmin==1.2.0
Rtree==1.1.0
s3fs==2023.10.0
s3transfer==0.7.0
shapely==2.0.2
six==1.16.0
sniffio==1.3.0
snuggs==1.4.7
soupsieve==2.5
SQLAlchemy==2.0.25
sqlparse==0.4.3
starlette==0.35.1
sympy==1.12
tenacity==8.2.3
tomli==2.0.1
torch==2.0.1
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
urllib3==1.26.13
uvicorn==0.27.0
uvloop==0.19.0
vine==5.0.0
virtualenv==20.25.0
warrant==0.6.1
watchfiles==0.21.0
wcwidth==0.2.6
websockets==12.0
whitenoise==6.6.0
wrapt==1.16.0
yarl==1.9.2
YubiOTP==1.0.0
zipp==3.15.0
